simple-stream: add to_input_stream()
Users of output stream may want to read what was just written. For
example a lower layer may give an output stream to a higher layer and
then want to compute a checksum of the data that the higher layer has
written. Such scenarios can be easily supported if input streams can be
created from output streams.