rpc: introduce lz4_fragmented_compressor
A serious flaw of lz4_compressor is that it needs to allocate contiguous
input and output buffers for the whole message. Apart from the cost of
linearisation this will also put additional pressure on the memory allocator
which may struggle to provide large contiguous buffers.

One solution may be to reuse those large buffers, but that means both
compressed and decompressed data would have to be copied to/from those
large buffers. That may not be an issue if the message is large because it
contains a lot of small elements, each adding a significant processing cost
later in the pipeline, but if the message is large merely because it contains
large blobs of opaque binary data then it is likely that the cost of copying
will dominate and each copy would add a measurable overhead.

The LZ4_FRAGMENTED compressor addresses those problem and while more complex
than the naive approach allows compression and decompression without the
need to allocate any large contiguous buffers. The compressed data is copied
once, the decompressed is either read directly form the source buffer (for
compression) or decompressed directly to the destination buffer
(for decompression). This should make this compressor a good choice for the
vast majority of uses.

Under the hood LZ4_FRAGMENTED uses lz4 stream inferface with custom metadata
(4 bytes per 128 kB of uncompressed data). This allows compressing data in
chunks without sacrificing compression ratio since compression of the current
chunk depends on the previous one (i.e. they are not independent from each
other and need to be decompressed in order, which is not a proble in this
case).