task: add dispose()
Currently, it is assumed that all task objects are separate allocations
on heap. This is not true since there are tasks like thread_wake_task
that are always allocated on stack. Such objects should not be handled
by std::unique_ptr<task>.

This is solved by adding a virtual function task::dispose() that
disposes a task in cases when it wasn't run (e.g. the promise was
dropped). Custom deleter for std::unique_ptr calls that function thus
making it safe to handle task objects regardless of their lifetime
situation provided dispose() is correctly implemented.

From the performance perspective there situation doesn't really change
since instead of virtual dtor we now have calls to a virtual function and
all concrete task types are marked final for devirtualisation.

Apart from the existing users, this change is going to be needed by
coroutines which store the task object as a part of their internal state
which lifetime is managed by the compiler.