add support for coroutines
This patch adds all necessary glue to make coroutines work with seastar
futures and promises. They integrate seamlessly and any function that
returns a future may become a coroutine.

seastar/core/coroutines.hh also defines some classes that should be in
experimental/coroutine (unless that file exists, in which case it is
included instead) so that this can work with libstdc++ which doesn't
provide that header.