rpc/compressor/lz4: drop broken special-case for uncompressed messages
LZ4 compressor decompression routine has a special case for situation
when the sender decided not to compress the message. Right, now it never
happens, but the idea was that in the future it might and the existing
receivers will already be able to receive such messages. Unfortunately,
that special case in the decompression code was unused, untested, and
broken since it didn't strip the metadata from the buffer returned to
the user.