core/posix_file_impl: enforce IOV_MAX on vectored IO
The size of the iovec array used by readv() and writev() is limited to
IOV_MAX. This means that most of the users of the vectored IO would have
to ensure that limit is not exceeded in order to avoid EINVAL.

This patch makes posix_file_impl::{read,write}_dma() enforce the limit by
trmming too long vectors of iovecs. This is okay, since the postconditions
alread allow partial operations. What it gives us is less preconditions,
which means safer to use interface. If the user code wants, it may still
handle cases where IOV_MAX would be exceeded on its own.